requirements-research.txt
# Research System Dependencies
# Install with: pip install -r requirements-research.txt

# GitHub API integration
PyGithub>=2.1.1

# YAML configuration parsing
PyYAML>=6.0.1

# Data analysis and processing
pandas>=2.0.0
numpy>=1.24.0

# Machine learning (optional, for future phases)
scikit-learn>=1.3.0

# Web scraping (for awesome lists, etc.)
beautifulsoup4>=4.12.0
requests>=2.31.0

# Caching
diskcache>=5.6.0

# Utilities
python-dateutil>=2.8.2
